Needed to add recursive support to mcopy, also other minor fixes

+++ mkbiosimg.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+#
+# PROG: Build a bootable disk image for PXE booting
+#
+
 set -e
 
 cd "$( dirname "${BASH_SOURCE[0]}" )" 
@@ -13,14 +17,31 @@ IMGNAME="`dirname \"$ZIPNAME\"`/`basename \"$ZIPNAME\" .zip`.img"
 
 FDOSNAME=mkbiosimg/fdboot.img
 
-/sbin/mkfs.vfat -C "$IMGNAME" 65536 
+#echo "==> MKFS.VFAT Create blank image"
+if [ ! -x /sbin/mkfs.vfat ]; then
+  echo " [!] ERROR: Cannot execute mkfs.vfat, exiting !"
+  exit 1
+else
+  /sbin/mkfs.vfat -C "$IMGNAME" 65536 
+fi
+
+#echo "==> SYS-FREEDOS: Copy FreeDOS boot sector info"
 mkbiosimg/sys-freedos.pl --disk="$IMGNAME"
+
+#echo "==>  7Z: Expand fdBoot images accordingly. FDOSNAME=${FDOSNAME} ZIPNAME=${ZIPNAME}"
+#echo "==> PWD: `pwd`"
 rm -rf tmp
 mkdir -p tmp/Update
 7z x -otmp "$FDOSNAME"
 7z x -otmp/Update "$ZIPNAME"
-mcopy -o -i"$IMGNAME" tmp/* ::
 
+#echo "==> MCOPY: Copy contents to a single directory IMGNAME ${IMGNAME} to tmp/\*"
+#echo "==>  PWD: `pwd`"
+# -v = verbose, -s = recursive, -o = overwrite, -i = image
+mcopy -s -o -i"$IMGNAME" tmp/*  ::
+
+#echo "==> PIGZ: Compress the image accordingly as a final product IMGNAME=${IMGNAME}"
+#echo "==>  PWD: `pwd`"
 pigz "$IMGNAME"
 ls -la "$IMGNAME.gz"
